% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{advected_variogram}
\alias{advected_variogram}
\title{advected_variogram function}
\usage{
advected_variogram(x, y, z, grid, model, adv)
}
\arguments{
\item{x}{Vector for the first dimension (spatial x in our case).}

\item{y}{Vector for the second dimension (spatial y in our case)}

\item{grid}{The grid matrix.}

\item{model}{The model to use for the variogram.}

\item{adv}{The advection coordinates vector. Default is c(0, 0).}

\item{t}{Vector for the third dimension (time in our case).}
}
\value{
An array for the spatio-temporal advected variogram.
}
\description{
This function computes the spatio-temporal advected variogram.
}
