% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{aggregate_chi_by_distance}
\alias{aggregate_chi_by_distance}
\title{Aggregate chi values by distance bins
This function aggregates chi values by distance bins.}
\usage{
aggregate_chi_by_distance(
  chi_table,
  nb_bins = 10,
  method = c("fixed", "quantile")
)
}
\arguments{
\item{chi_table}{The data frame containing chi values and distances.}

\item{nb_bins}{The number of bins to use for aggregation.}

\item{method}{The method to use for binning: "fixed" or "quantile".}
}
\value{
A data frame containing the aggregated chi values by distance bins.
}
\description{
Aggregate chi values by distance bins
This function aggregates chi values by distance bins.
}
