% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{Calculate the root mean squared error (RMSE) between true values and
predicted values.}
\usage{
calculate_rmse(true_values, predicted_values)
}
\arguments{
\item{true_values}{The vector of true values.}

\item{predicted_values}{The vector of predicted values.}
}
\value{
The root mean squared error (RMSE) between true values and
        predicted values.
}
\description{
Calculate the root mean squared error (RMSE) between true values and
predicted values.
}
