% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{cardinal_to_degree}
\alias{cardinal_to_degree}
\title{cardinal_to_degree function}
\usage{
cardinal_to_degree(cardinal)
}
\arguments{
\item{cardinal}{The cardinal direction.}
}
\value{
The wind direction in degrees.
}
\description{
Convert cardinal direction between N, NE, E, SE, S, SW, W, NW to degrees.
}
