% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{check_intervals_overlap}
\alias{check_intervals_overlap}
\title{check_intervals_overlap function}
\usage{
check_intervals_overlap(s0_value, selected_points, delta, beta = 0)
}
\arguments{
\item{s0_value}{The s0 value.}

\item{selected_points}{The selected points (s0, t0, u_s0).}

\item{delta}{The temporal window size.}

\item{beta}{The temporal window extension. Default is 0.}
}
\value{
TRUE if the intervals overlap, FALSE otherwise.
}
\description{
This function checks if the time intervals overlap for all extreme episodes
associated with a given s0 value.
}
