% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{chi_vario}
\alias{chi_vario}
\title{Calculate the chi-squared variogram}
\usage{
chi_vario(vario)
}
\arguments{
\item{vario}{The variogram value.}
}
\value{
The theorical spatio-temporal extremogram.
}
\description{
This function calculates the chi-squared variogram based on the given
variogram value.
}
\examples{
vario <- 0.5
chi_vario(vario)

}
