% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{chispatemp_empirical}
\alias{chispatemp_empirical}
\title{chispatemp_empirical function}
\usage{
chispatemp_empirical(data_rain, df_lags, quantile, remove_zeros = TRUE)
}
\arguments{
\item{data_rain}{The data set.}

\item{df_lags}{The dataframe containing the spatial and temporal lags.}

\item{quantile}{The quantile value.}
}
\value{
The spatio-temporal chi value.
}
\description{
Calculate the spatio-temporal chi extremogram for a given dataset
and a quantile value.
}
