% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{compute_all_pairs_chi}
\alias{compute_all_pairs_chi}
\title{Compute chi for all pairs of sites
This function computes the chi statistic for all pairs of sites
in the dataset.}
\usage{
compute_all_pairs_chi(data_rain, dist_mat, q = 0.99, zeros = TRUE)
}
\arguments{
\item{data_rain}{The rainfall data.}

\item{dist_mat}{The distance matrix.}

\item{q}{The quantile value.}

\item{zeros}{Logical indicating whether to include zero values in the
calculation.}
}
\value{
A data frame containing the chi values for all pairs of sites.
}
\description{
Compute chi for all pairs of sites
This function computes the chi statistic for all pairs of sites
in the dataset.
}
