% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{compute_gamma_point}
\alias{compute_gamma_point}
\title{compute_gamma_point function}
\usage{
compute_gamma_point(grid, gamma_space, gamma_time, adv, s = NA, t = NA)
}
\arguments{
\item{grid}{The grid matrix.}

\item{gamma_space}{The spatial variogram.}

\item{gamma_time}{The temporal variogram.}

\item{adv}{The advection coordinates vector.}

\item{s}{The spatial coordinates vector. Default is NA.}

\item{t}{The temporal time. Default is NA.}
}
\value{
The spatio-temporal variogram for the given point.
}
\description{
This function computes the spatio-temporal variogram for a given point.
}
