% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{compute_p0_all_episodes}
\alias{compute_p0_all_episodes}
\title{compute_p0_all_episodes}
\usage{
compute_p0_all_episodes(list_episodes)
}
\arguments{
\item{list_episodes}{list of matrices of observed values (sites x time)}
}
\value{
list with matrix of p0 values by episode and mean p0 values
}
\description{
Compute the probability of zero for each site across multiple episodes
}
