% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{compute_p0_episode}
\alias{compute_p0_episode}
\title{compute_p0_episode}
\usage{
compute_p0_episode(episode)
}
\arguments{
\item{episode}{matrix of observed values (sites x time)}
}
\value{
named numeric vector of p0 values for each site
}
\description{
Compute the probability of zero for each site in a single episode
}
