% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd.R
\name{compute_rmse}
\alias{compute_rmse}
\title{compute_rmse function
This function computes the Root Mean Square Error (RMSE) between empirical
and theoretical quantiles of the Extreme Generalized Pareto Distribution (EGPD)
for a given dataset and left censoring value.}
\usage{
compute_rmse(y, left_censore)
}
\arguments{
\item{y}{The data to be evaluated.}

\item{left_censore}{The left censoring value.}
}
\value{
The computed RMSE value.
}
\description{
compute_rmse function
This function computes the Root Mean Square Error (RMSE) between empirical
and theoretical quantiles of the Extreme Generalized Pareto Distribution (EGPD)
for a given dataset and left censoring value.
}
