% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{compute_st_gaussian_process}
\alias{compute_st_gaussian_process}
\title{compute_st_gaussian_process function}
\usage{
compute_st_gaussian_process(
  grid,
  W_s = NULL,
  W_s_x = NULL,
  W_s_y = NULL,
  W_t,
  adv
)
}
\arguments{
\item{grid}{The grid matrix.}

\item{W_s}{The spatial Gaussian random field (for euclidean case).}

\item{W_s_x}{The spatial Gaussian random field in the x direction (for lalpha case).}

\item{W_s_y}{The spatial Gaussian random field in the y direction (for lalpha case).}

\item{W_t}{The temporal Gaussian random field.}

\item{adv}{The advection coordinates vector.}
}
\value{
The spatio-temporal Gaussian random field.
}
\description{
This function computes the spatio-temporal Gaussian random field given the
spatial and temporal Gaussian random fields and the advection.
}
