% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{compute_st_variogram}
\alias{compute_st_variogram}
\title{compute_st_variogram function}
\usage{
compute_st_variogram(
  grid,
  gamma_space = NULL,
  gamma_space_x = NULL,
  gamma_space_y = NULL,
  gamma_temp,
  adv
)
}
\arguments{
\item{grid}{The grid data.frame with columns x, y, t, shifted_x, shifted_y.}

\item{gamma_space}{The isotropic spatial variogram (optional).}

\item{gamma_space_x}{The spatial variogram in the x direction (optional).}

\item{gamma_space_y}{The spatial variogram in the y direction (optional).}

\item{gamma_temp}{The temporal variogram.}

\item{adv}{The advection vector (length 2 numeric).}
}
\value{
A 3D array with the spatio-temporal variogram values over the grid.
}
\description{
Computes the spatio-temporal variogram over a grid.
}
