% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{compute_W_s_t}
\alias{compute_W_s_t}
\title{compute_W_s_t function}
\usage{
compute_W_s_t(grid, W_s, W_t, adv)
}
\arguments{
\item{grid}{The grid matrix.}

\item{W_s}{The spatial gaussian random field.}

\item{W_t}{The temporal gaussian random field.}

\item{adv}{The advection coordinates vector.}
}
\value{
The spatio-temporal gaussian random field.
}
\description{
This function computes the spatio-temporal gaussian random field given the
spatial and temporal gaussian random fields and the advection.
}
