% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{conditional_empirical_excesses}
\alias{conditional_empirical_excesses}
\title{conditional_empirical_excesses function}
\usage{
conditional_empirical_excesses(
  data_rain,
  quantile,
  df_lags,
  s0 = c(1, 1),
  t0 = 1
)
}
\arguments{
\item{data_rain}{The rainfall data.}

\item{quantile}{The quantile value.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{s0}{Spatial location vector for conditional point}

\item{t0}{Conditional time}
}
\value{
The conditional empirical excesses dataframe with the number of
excesses kij and the number of possible excesses Tobs, with the lag values.
}
\description{
This function calculates the empirical excesses based on indicators above a
quantile threshold conditioning on a spatio-temporal point (s0, t0)
}
