% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{conditional_variogram}
\alias{conditional_variogram}
\title{conditional_variogram function}
\usage{
conditional_variogram(x, y, t, s0, t0, grid, model, adv = c(0, 0))
}
\arguments{
\item{x}{Vector for the first dimension (spatial x in our case).}

\item{y}{Vector for the second dimension (spatial y in our case)}

\item{t}{Vector for the third dimension (time in our case).}

\item{s0}{Vector of dimension 2 for the spatial conditioning point.}

\item{t0}{Value of the temporal conditioning point.}

\item{grid}{The grid matrix.}

\item{model}{The model to use for the variogram.}

\item{adv}{The advection coordinates vector. Default is c(0, 0).}
}
\value{
The spatio-temporal conditional variogram.
}
\description{
This function computes the spatio-temporal conditional variogram.
}
