% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{convert_params}
\alias{convert_params}
\title{convert_params function}
\usage{
convert_params(beta1, beta2, alpha1, alpha2, c_x = 1, c_t = 1)
}
\arguments{
\item{beta1}{The spatial scale parameter.}

\item{beta2}{The temporal scale parameter.}

\item{alpha1}{The spatial smoothness parameter.}

\item{alpha2}{The temporal smoothness parameter.}

\item{c_x}{The spatial scaling constant. Default is 1.}

\item{c_t}{The temporal scaling constant. Default is 1.}
}
\value{
A list containing the converted parameters beta1 and beta2.
}
\description{
Convert variogram parameters based on scaling constants for space and time.
}
