% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{convert_simulations_to_list}
\alias{convert_simulations_to_list}
\title{Convert simulations from array to list of data.frames}
\usage{
convert_simulations_to_list(simu, sites_coords)
}
\arguments{
\item{simu}{The simulation array (x, y, t, nsimu)}

\item{ngrid}{Size of spatial grid (used to build coordinates)}
}
\value{
A list of data.frames, one per episode
}
\description{
Convert simulations from array to list of data.frames
}
