% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{convert_single_simulation_to_df}
\alias{convert_single_simulation_to_df}
\title{Convert a single simulation (3D array) to a data.frame}
\usage{
convert_single_simulation_to_df(simu_i, ngrid)
}
\arguments{
\item{simu_i}{A 3D array of dimensions (x, y, t) corresponding to one simulation}

\item{ngrid}{The size of the spatial grid}
}
\value{
A data.frame with each column being a spatial point and rows as time steps
}
\description{
Convert a single simulation (3D array) to a data.frame
}
