% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{convert_to_cardinal}
\alias{convert_to_cardinal}
\title{convert_to_cardinal function}
\usage{
convert_to_cardinal(degrees)
}
\arguments{
\item{degrees}{The wind direction in degrees.}
}
\value{
The cardinal direction.
}
\description{
Convert wind direction in degrees to cardinal direction between N, NE, E, SE,
S, SW, W, NW.
}
