% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_simu_gif}
\alias{create_simu_gif}
\title{Create a GIF from a simulation}
\usage{
create_simu_gif(
  simulation_data,
  sites_coords,
  params,
  foldername,
  type = "rpar",
  forcedtemp = NA,
  s0 = NULL,
  threshold = 1
)
}
\arguments{
\item{simulation_data}{The data from the simulation.}

\item{sites_coords}{The coordinates of the sites.}

\item{params}{The parameters used for the simulation.}

\item{foldername}{The folder where the GIF will be saved.}

\item{type}{The type of simulation.}

\item{forcedtemp}{The number of time steps to include in the GIF.}

\item{s0}{The conditional site.}

\item{threshold}{The threshold value to indicate on the plot. Default is 1.}
}
\value{
None, but saves the GIF to a file.
}
\description{
This function creates a GIF from a simulation.
}
