% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{dchi_dalpha1}
\alias{dchi_dalpha1}
\title{Compute the derivative of the extremogram with respect to alpha1.}
\usage{
dchi_dalpha1(params, hnorm, tau)
}
\arguments{
\item{params}{a vector of variogram parameters.}

\item{hnorm}{a vector of spatial lags.}

\item{tau}{a vector of temporal lags.}
}
\value{
The derivative of the extremogram with respect to alpha1.
}
\description{
This function calculates the derivative of theextremogram with respect to
alpha1.
}
\examples{
params <- c(0.5, 0.8)
hnorm <- 0.2
tau <- 0.1
dchi_dalpha1(params, hnorm, tau)

}
