% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{dchi_dalpha2}
\alias{dchi_dalpha2}
\title{Compute the derivative of the extremogram with respect to alpha2.}
\usage{
dchi_dalpha2(params, hnorm, tau)
}
\arguments{
\item{params}{a vector of variogram parameters.}

\item{hnorm}{a vector of spatial lags.}

\item{tau}{a vector of temporal lags.}
}
\value{
The derivative of the extremogram with respect to alpha2
}
\description{
This function calculates the derivative of the extremogram with respect
to alpha2.
}
\examples{
params <- c(0.5, 0.2)
hnorm <- c(0.1, 0.3, 0.2, 0.4)
tau <- 0.5
dchi_dalpha2(params, hnorm, tau)

}
