% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{dchi_dbeta1}
\alias{dchi_dbeta1}
\title{Compute the derivative of the extremogram with respect to beta1.}
\usage{
dchi_dbeta1(params, hnorm, tau)
}
\arguments{
\item{params}{a vector of variogram parameters.}

\item{hnorm}{a vector of spatial lags.}

\item{tau}{a vector of temporal lags.}
}
\value{
The derivative of the extremogram with respect to beta1.
}
\description{
This function calculates the derivative of the extremogram with respect to
beta1.
}
\examples{
dchi_dbeta1(params = c(1, 2, 3), hnorm = 0.5, tau = 0.1)

}
