% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{dchi_dbeta2}
\alias{dchi_dbeta2}
\title{Compute the derivative of the extremogram function with respect to beta2.}
\usage{
dchi_dbeta2(params, hnorm, tau)
}
\arguments{
\item{params}{a vector of variogram parameters.}

\item{hnorm}{a vector of spatial lags.}

\item{tau}{a vector of temporal lags.}
}
\value{
The derivative of the extremogram function with respect to beta2.
}
\description{
This function calculates the derivative of the extremogram function with
respect to beta2.
}
\examples{
params <- c(1, 2)
hnorm <- c(0.1, 0.2, 0.3, 0.4)
tau <- 0.5
dchi_dbeta2(params, hnorm, tau)

}
