% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd.R
\name{dgpdExt1}
\alias{dgpdExt1}
\title{dgpdExt1 function}
\usage{
dgpdExt1(x, kappa, sigma, gamma)
}
\arguments{
\item{x}{The input value.}

\item{kappa}{The kappa parameter (power transform).}

\item{sigma}{The sigma parameter (scale).}

\item{gamma}{The gamma parameter (shape).}
}
\value{
The calculated dgpdExt1 value.
}
\description{
This function calculates the density distribution of the EGPD with the first 
model $G(v) = v^\kappa$ based on the given parameters.
}
