% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{dist_adv}
\alias{dist_adv}
\title{dist_adv function}
\usage{
dist_adv(s1, s2, t1, t2, adv)
}
\arguments{
\item{s1}{The first point coordinates.}

\item{s2}{The second point coordinates.}

\item{t1}{The first time.}

\item{t2}{The second time.}

\item{adv}{The advection coordinates vector.}
}
\value{
The distance between the two points with advection.
}
\description{
This function computes the distance between two points with advection.
}
