% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances_regular_grid}
\alias{distances_regular_grid}
\title{distances_regular_grid calculates the distances between sites on a regular
grid.}
\usage{
distances_regular_grid(nsites)
}
\arguments{
\item{nsites}{the number of sites on the grid ie nb of pixels^2}
}
\value{
a matrix of distances between sites
}
\description{
distances_regular_grid calculates the distances between sites on a regular
grid.
}
\examples{
distances_regular_grid(10)

}
