% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{empirical_excesses}
\alias{empirical_excesses}
\title{empirical_excesses function}
\usage{
empirical_excesses(data_rain, quantile, df_lags, threshold = FALSE)
}
\arguments{
\item{data_rain}{The rainfall data.}

\item{quantile}{The quantile value.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{threshold}{A boolean value to indicate if the quantile variable is a
threshold value and not a uniform quantile. Default is FALSE.}
}
\value{
The empirical excesses dataframe with the number of excesses kij and
the number of possible excesses Tobs, with the lag values.
}
\description{
This function calculates the empirical excesses based on indicators above a
quantile threshold.
}
