% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{eta}
\alias{eta}
\title{Transforms the extremogram using the eta transformation.}
\usage{
eta(chi)
}
\arguments{
\item{chi}{The input extremogram to be transformed.}
}
\value{
The transformed vector.
}
\description{
The eta function transforms the input vector by applying a logarithmic
transformation using the quantile function of the standard normal
distribution.
}
\examples{
eta(c(-1, 0, 1, 2))
# Output: [1]  2.000000e+00  1.000000e-06 -1.151293e-01 -2.000000e-01

}
