% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{evaluate_optim}
\alias{evaluate_optim}
\title{evaluate_optim function}
\usage{
evaluate_optim(list_simu, quantile, true_param, tau, df_dist)
}
\arguments{
\item{list_simu}{A list of simulated data}

\item{quantile}{The quantile value}

\item{true_param}{The true variogram parameter (beta1, beta2, alpha1, alpha2)}

\item{tau}{The temporal lag vector}
}
\value{
The result of the optimization process as a dataframe.
}
\description{
This function evaluates the optimization process on simulated data (for
example, following a Brown-Resnick process) using the negative
log-likelihood function.
}
\examples{
evaluate_optim(list_simu, quantile, true_param, tau)
}
