% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{evaluate_optim_simuExp}
\alias{evaluate_optim_simuExp}
\title{Evaluate optimization using simulated binomial experiments}
\usage{
evaluate_optim_simuExp(n_res, Tmax, tau_vect, h_vect, chi, df_dist, nconfig)
}
\arguments{
\item{n_res}{The number of realizations}

\item{Tmax}{The maximum time observation}

\item{tau_vect}{The temporal lag vector}

\item{h_vect}{The spatial lag vector}

\item{chi}{The spatio-temporal extremogram matrix}

\item{df_dist}{The distances long dataframe}

\item{nconfig}{The number of configurations}
}
\value{
The result of the optimization process as a dataframe.
}
\description{
This function evaluates the optimization process using simulated experiments.
Simulation of excesses is performed considering a binomial distribution with
a probability parameter equals to the spatio-temporal extremogram value.
}
\examples{
evaluate_optim_simuExp(10, 100, 1:10, 1:10, chi_mat, df_dist)
}
