% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{evaluate_vario_estimates}
\alias{evaluate_vario_estimates}
\title{Evaluate variogram estimates}
\usage{
evaluate_vario_estimates(
  list_simu,
  quantile,
  spatial = TRUE,
  df_dist = NA,
  tmax = NA,
  hmax = NA
)
}
\arguments{
\item{list_simu}{The list of dataframes containing the simulations.}

\item{quantile}{The quantile for the chiplot.}

\item{spatial}{Logical indicating whether to use spatial chi.}

\item{df_dist}{The dataframe containing the distances between the sites for
spatial chi.}

\item{tmax}{The maximum temporal lag for temporal chi.}

\item{hmax}{The maximum spatial lag for spatial chi.}
}
\value{
A list containing the dataframe of valid results and the results.
}
\description{
This function evaluates variogram estimates over all simulations and returns
the mean, the RMSE and the MAE of the estimates.
}
