% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd.R
\name{fit_egpd_site}
\alias{fit_egpd_site}
\title{fit_egpd_site function
This function fits the Extreme Generalized Pareto Distribution (EGPD)
to the given data for a specific site.}
\usage{
fit_egpd_site(y, site_name, censures = seq(0.22, 0.26, by = 0.01), R = 1000)
}
\arguments{
\item{y}{The data to be fitted.}

\item{site_name}{The name of the site}

\item{censures}{The vector of censures to be tested. Default is seq(0.22, 0.26, by = 0.01)}

\item{R}{The number of bootstrap samples for confidence intervals. Default is 1000}
}
\value{
A data frame containing the fitted parameters and best censore for the site.
}
\description{
fit_egpd_site function
This function fits the Extreme Generalized Pareto Distribution (EGPD)
to the given data for a specific site.
}
