% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_estimate}
\alias{format_estimate}
\title{Formats p-values with asterisks}
\usage{
format_estimate(estimate, p_value)
}
\arguments{
\item{estimate}{A numeric vector of coefficient estimates.}

\item{p_value}{A numeric vector of p-values corresponding to the
coefficient estimates.}
}
\value{
A character vector of formatted coefficient estimates
         with asterisks indicating significance.
}
\description{
This function formats coefficient estimates by adding asterisks
to indicate significance.
}
