% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_value}
\alias{format_value}
\title{Format Numeric Values into Custom String Representation}
\usage{
format_value(x)
}
\arguments{
\item{x}{A numeric vector containing the values to be formatted.}
}
\value{
A single string where each formatted value is concatenated with
underscores. The formatting rules are as follows:
- If the value is negative, the string "neg" is prefixed to the formatted
  value.
- If the value is an integer, it is formatted as is.
- If the value is a decimal:
  - For values greater than or equal to 1:
    - If there is 1 decimal place, the value is multiplied by 10 and
      formatted with 2 digits.
    - If there are more than 1 decimal places, the value is multiplied by 100
      and formatted with 3 digits.
  - For values less than 1:
    - If there is 1 decimal place, the value is multiplied by 10 and
      formatted with 2 digits.
    - If there are more than 1 decimal places, the value is multiplied by 100
      and formatted with 3 digits.
}
\description{
This function takes a numeric vector and formats each value into a custom
string representation. The formatted values are concatenated with
underscores.
}
\examples{
format_value(c(-1.23, 0.5, 2, -0.01))
# Returns: "neg123_05_200_neg001"

}
