% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{G_std}
\alias{G_std}
\title{G_std}
\usage{
G_std(z, p0, u)
}
\arguments{
\item{z}{numeric vector of latent variables}

\item{p0}{probability of zero}

\item{u}{threshold in the latent space}
}
\value{
numeric vector of transformed variables
}
\description{
Standardized CDF transformation function to get rainfall scale from latent 
pareto scale
}
