% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{G_std_inv}
\alias{G_std_inv}
\title{G_std_inv}
\usage{
G_std_inv(v, p0, u)
}
\arguments{
\item{v}{numeric vector of transformed variables}

\item{p0}{probability of zero}

\item{u}{threshold in the latent space}
}
\value{
numeric vector of latent variables
}
\description{
Inverse of the standardized CDF transformation function to get latent 
pareto scale from rainfall scale
}
