% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gamma_theta}
\alias{gamma_theta}
\title{vario function}
\usage{
gamma_theta(h, tau, beta1, beta2, alpha1, alpha2)
}
\arguments{
\item{h}{The spatial lags.}

\item{tau}{The temporal lags.}

\item{beta1}{The value of beta1.}

\item{beta2}{The value of beta2.}

\item{alpha1}{The value of alpha1.}

\item{alpha2}{The value of alpha2.}
}
\value{
The variogram value.
}
\description{
This function computes the variogram value for a given distance matrix and
variogram parameters.
}
