% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_data_for_tau}
\alias{generate_data_for_tau}
\title{Generate data for a specific tau value for the variogram plot}
\usage{
generate_data_for_tau(tau, empirical_df, theorical_df)
}
\arguments{
\item{tau}{The value of tau to generate data for.}

\item{empirical_df}{The dataframe of empirical variogram values.}

\item{theorical_df}{The dataframe of theoretical variogram values.}
}
\value{
A dataframe containing the variogram values for the specified tau.
}
\description{
This function generates data for a specific tau value for the variogram plot.
}
