% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{generate_grid_coords}
\alias{generate_grid_coords}
\title{generate_grid_coords generates the spatial coordinates of a regular grid.}
\usage{
generate_grid_coords(grid_size)
}
\arguments{
\item{grid_size}{the size of the grid (number of rows/columns)}
}
\value{
a data frame containing the grid coordinates
}
\description{
generate_grid_coords generates the spatial coordinates of a regular grid.
}
\examples{
generate_grid_coords(5)

}
