% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{generate_realistic_latlon_grid}
\alias{generate_realistic_latlon_grid}
\title{generate_realistic_latlon_grid generates a data frame with site IDs and
random latitude and longitude values within given ranges.}
\usage{
generate_realistic_latlon_grid(
  n_sites,
  lat_range = c(40, 50),
  lon_range = c(-5, 5)
)
}
\arguments{
\item{n_sites}{the number of sites to generate}

\item{lat_range}{the range of latitude values. Default is c(40, 50).}

\item{lon_range}{the range of longitude values. Default is c(-5, 5).}
}
\value{
a data frame with site IDs and coordinates
}
\description{
generate_realistic_latlon_grid generates a data frame with site IDs and
random latitude and longitude values within given ranges.
}
