% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_variogram_plots_rpareto}
\alias{generate_variogram_plots_rpareto}
\title{Generate a variogram plot for rpareto}
\usage{
generate_variogram_plots_rpareto(result, df_lags, wind_df)
}
\arguments{
\item{result}{The result of the variogram estimation.}

\item{df_lags}{The dataframe of lags.}

\item{wind_df}{The dataframe of wind data.}

\item{tau_values}{The values of tau to generate plots for.}

\item{chi}{A logical value indicating whether to plot the chi values.
Default is FALSE.}
}
\value{
None, but generates plots.
}
\description{
This function generates a variogram plot for the rpareto model.
}
