% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_chi_vect}
\alias{get_chi_vect}
\title{get_chi_vect function}
\usage{
get_chi_vect(chi_mat, h_vect, tau, df_dist)
}
\arguments{
\item{chi_mat}{The theoretical spatio-temporal extremogram matrix.}

\item{h_vect}{The spatial lags vector.}

\item{tau}{The temporal lag vector.}

\item{df_dist}{The distances long dataframe.}
}
\value{
The theoretical spatio-temporal extremogram vector.
}
\description{
This function calculates the theoretical spatio-temporal extremogram vector
based on the given theoretical spatio-temporal extremogram matrix.
}
