% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_criterion}
\alias{get_criterion}
\title{get_criterion function}
\usage{
get_criterion(df_result, true_param)
}
\arguments{
\item{df_result}{The data frame containing the result values}

\item{true_param}{The true variogram parameter (beta1, beta2, alpha1, alpha2)}
}
\value{
The calculated criterion value in a dataframe.
}
\description{
Calculates the mean, RMSE and MAE criterion values for the given dataframe
result and the true parameter values.
}
