% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd.R
\name{get_df_long_params_egpd}
\alias{get_df_long_params_egpd}
\title{get_df_long_params_egpd function}
\usage{
get_df_long_params_egpd(df_params)
}
\arguments{
\item{df_params}{A dataframe containing parameters.}
}
\value{
A long format dataframe.
}
\description{
This function takes a dataframe of parameters and returns a long format
dataframe for EGPD estimates.
}
