% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_dist_mat}
\alias{get_dist_mat}
\title{This function reshapes a distance matrix into a long dataframe.}
\usage{
get_dist_mat(
  locations,
  dmax = NA,
  latlon = TRUE,
  adv = c(0, 0),
  tau = 1:10,
  alpha_spa = 1.5
)
}
\arguments{
\item{locations}{A matrix or data frame containing the coordinates of the
locations.}

\item{dmax}{The maximum distance threshold if specified.}

\item{latlon}{Logical indicating whether the coordinates are in latitude and
longitude format. Default is TRUE.}

\item{adv}{A vector of advection values. Default is c(0, 0).}

\item{tau}{A vector of temporal lags. Default is 1:10.}

\item{alpha_spa}{The spatial lag exponent. Default is 1.5.}
}
\value{
A distance matrix containing the reshaped distances.
}
\description{
This function reshapes a distance matrix into a long dataframe.
}
