% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd.R
\name{get_egpd_estimates}
\alias{get_egpd_estimates}
\title{get_egpd_estimates function}
\usage{
get_egpd_estimates(rain_df, left_censoring = 0)
}
\arguments{
\item{rain_df}{A data frame containing the rainfall data.}

\item{left_censoring}{The threshold for left censoring. Default is 0.}
}
\value{
A list containing the estimates of the EGPD parameters.
}
\description{
This function calculates the estimates of the Extreme Generalized Pareto
Distribution (EGPD) parameters.
}
\examples{
data <- read.csv("rainfall_data.csv")
estimates <- get_egpd_estimates(data, left_censoring = 10)

}
