% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{get_estimate_variospa}
\alias{get_estimate_variospa}
\title{Calculate the estimate of the spatial variogram parameters}
\usage{
get_estimate_variospa(chispa, weights, summary = FALSE)
}
\arguments{
\item{chispa}{The spatial chi values.}

\item{weights}{The weights to be used in the calculation: "residuals", "exp",
or "none"}

\item{summary}{Logical value indicating whether to return the summary of the
model. Default is FALSE.}
}
\value{
The estimated spatial variogram parameters.
}
\description{
This function calculates the estimate of the spatial variogram parameters
based on the given spatial chi values. It uses the Weighted Least Squares
Estimation (WLSE) method to estimate the parameters.
}
\examples{
chispa <- data.frame(chi = c(0.5, 0.6, 0.7), lagspa = c(1, 2, 3))
weights <- "residuals"
get_estimate_variospa(chispa, weights)
}
