% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_euclidean_distance}
\alias{get_euclidean_distance}
\title{Calculate the Euclidean distance between two points.}
\usage{
get_euclidean_distance(point1, point2)
}
\arguments{
\item{point1}{The first point.}

\item{point2}{The second point.}
}
\value{
The Euclidean distance between the two points.
}
\description{
Calculate the Euclidean distance between two points.
}
\examples{
euclidean_distance(c(0, 0), c(3, 4))

}
