% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{get_extreme_episodes}
\alias{get_extreme_episodes}
\title{get_extreme_episodes function}
\usage{
get_extreme_episodes(selected_points, data, delta, beta = 0)
}
\arguments{
\item{selected_points}{The selected points conditionning on X(s0,t0) > u.}

\item{data}{The data.}

\item{delta}{The episode length.}

\item{beta}{The buffer size. Default is 0.}
}
\value{
The list of extreme episodes.
}
\description{
Get the extreme episodes based on the selected points and the data by keeping
all sites for each episode.
}
