% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_h_vect}
\alias{get_h_vect}
\title{get_h_vect function}
\usage{
get_h_vect(df_dist, hmax)
}
\arguments{
\item{df_dist}{The distance dataframe.}

\item{hmax}{The maximum spacial lag value.}

\item{vectors}{A logical value indicating whether to return the spatial
lag vectors. Default is FALSE for only distance lags (norm).}
}
\value{
The spatial distance lags vector or list of lag vectors.
}
\description{
This function calculates the spatial lag vector based on the given distance
dataframe and maximum spacial lag value.
}
