% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_lag_vectors}
\alias{get_lag_vectors}
\title{get_lag_vectors function}
\usage{
get_lag_vectors(df_coords, hmax = NA, tau_max = 10)
}
\arguments{
\item{df_coords}{A dataframe containing the coordinates of the points.}

\item{hmax}{The maximum distance threshold. Default is NA.}

\item{tau_max}{The maximum temporal lag. Default is 10.}
}
\value{
A dataframe containing the lag vectors.
}
\description{
This function calculates the lag vectors between pairs of points.
}
