% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_marginal_excess}
\alias{get_marginal_excess}
\title{get_marginal_excess function}
\usage{
get_marginal_excess(
  data_rain,
  quantile,
  threshold = FALSE,
  index = 1,
  time = 1
)
}
\arguments{
\item{data_rain}{The rainfall data.}

\item{quantile}{The quantile value.}

\item{threshold}{A boolean value to indicate if the quantile variable is a
threshold value and not a uniform quantile. Default is FALSE.}

\item{index}{The index of the site. Default is 1.}

\item{time}{The starting time index. Default is 1.}
}
\value{
The number of marginal excesses.
}
\description{
This function calculates the number of marginal excesses above a given
quantile threshold.
}
