% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_mode_dir}
\alias{get_mode_dir}
\title{get_mode_dir function}
\usage{
get_mode_dir(x)
}
\arguments{
\item{x}{The wind data}
}
\value{
The most frequent wind direction.
}
\description{
Get the most frequent wind direction in a vector of wind directions.
}
