% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_results_optim}
\alias{get_results_optim}
\title{get_results_optim function}
\usage{
get_results_optim(filename, data_folder = NA)
}
\arguments{
\item{filename}{The name of the file to get the results.}

\item{data_folder}{The folder where the results are saved. Default is NA.}
}
\description{
Get the results of the optimization process from a CSV file.
}
