% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_s0t0_pairs}
\alias{get_s0t0_pairs}
\title{get_s0t0_pairs function}
\usage{
get_s0t0_pairs(
  sites_coords,
  data,
  min_spatial_dist,
  episode_size,
  set_st_excess,
  n_max_episodes = 10000,
  latlon = FALSE,
  beta = 0
)
}
\arguments{
\item{sites_coords}{The coordinates of the sites.}

\item{data}{The rainfall data as a matrix.}

\item{min_spatial_dist}{The minimum spatial distance between two episodes.}

\item{episode_size}{The temporal window size.}

\item{set_st_excess}{A list containing the spatio-temporal excesses with
list_s, list_t, and list_u.}

\item{n_max_episodes}{The maximum number of episodes to select. Default is 10000.}

\item{latlon}{A boolean value indicating if the coordinates are in latitude/longitude.}

\item{beta}{The temporal window extension. Default is 0.}
}
\value{
A data.table with selected pairs of sites and times (s0, t0) and their
        corresponding excess values u_s0.
}
\description{
This function retrieves pairs of sites and times (s0, t0) based on spatio-temporal
excesses. It ensures that selected pairs are sufficiently distant in both space and time.
}
