% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_spatiotemp_excess}
\alias{get_spatiotemp_excess}
\title{get_spatiotemp_excess function}
\usage{
get_spatiotemp_excess(
  data,
  quantile = NULL,
  threshold = NULL,
  remove_zeros = FALSE
)
}
\arguments{
\item{data}{The rainfall data as a matrix.}

\item{quantile}{The quantile value to compute the threshold.}

\item{threshold}{A fixed threshold value or vector.}

\item{remove_zeros}{A boolean indicating whether to remove zeros when computing quantiles.}
}
\value{
A list containing three vectors: list_s (sites), list_t (times),
}
\description{
This function computes the spatio-temporal excesses based on a quantile or
threshold. It returns a list of sites, times, and thresholds where the
excesses occur.
}
