% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{grad_neg_ll}
\alias{grad_neg_ll}
\title{Compute the gradient of the negative log-likelihood function}
\usage{
grad_neg_ll(
  params,
  simu,
  h_vect,
  tau,
  df_dist,
  locations,
  latlon = FALSE,
  quantile = 0.9,
  nmin = 5,
  simu_exp = FALSE,
  excesses = NULL
)
}
\arguments{
\item{params}{a vector of variogram parameters}

\item{simu}{a dataframe of simulated data}

\item{h_vect}{a vector of spatial lag values}

\item{tau}{a vector of temporal lag values}

\item{df_dist}{a numeric value representing the degrees of freedom for the
distribution}

\item{locations}{a matrix of locations}
}
\value{
a vector representing the gradient of the negative
        log-likelihood function
}
\description{
This function computes the gradient of the negative log-likelihood function
given a set of parameters, simulated data, bandwidth vector, tau value,
degrees of freedom for the distribution, and locations.
}
