% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{mae}
\alias{mae}
\title{Calculate the mean absolute error (MAE) between true values and
predicted values.}
\usage{
mae(true_values, predicted_values)
}
\arguments{
\item{true_values}{A numeric vector of true values.}

\item{predicted_values}{A numeric vector of predicted values.}
}
\value{
The mean absolute error between true values and predicted values.
}
\description{
Calculate the mean absolute error (MAE) between true values and
predicted values.
}
