% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{make_qq_df}
\alias{make_qq_df}
\title{make_qq_df}
\usage{
make_qq_df(x_obs, x_sim, min_n = 50)
}
\arguments{
\item{x_obs}{numeric vector of observed values}

\item{x_sim}{numeric vector of simulated values}

\item{min_n}{minimum number of values required to create the QQ data frame}
}
\value{
data frame with quantiles of observed and simulated values
}
\description{
Create a data frame for QQ plot from observed and simulated values
}
