% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll}
\alias{neg_ll}
\title{neg_ll function}
\usage{
neg_ll(
  params,
  simu,
  h_vect,
  tau,
  locations,
  latlon = FALSE,
  quantile = 0.9,
  nmin = 5,
  simu_exp = FALSE,
  excesses = NULL,
  adv1 = 0,
  adv2 = 0
)
}
\arguments{
\item{params}{Vector of variogram parameters (beta1, beta2, alpha1, alpha2).}

\item{h_vect}{The spatial lag vector.}

\item{tau}{The temporal lag vector.}

\item{locations}{The locations dataframe.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is FALSE.}

\item{simu_exp}{A boolean value to indicate if the data is simulated with
an exponential distribution.}

\item{excesses}{List of excesses (N_vect, n_vect).}

\item{df_dist}{The distances long dataframe.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate the negative log-likelihood for a given set of variogram
parameters by considering excess indicators following a binomial distribution
with a probability parameter equals to the theorical spatio-temporal chi
value. The number observations is different for each spatio-temporal
configuration.
}
