% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll_composite}
\alias{neg_ll_composite}
\title{neg_ll_composite function}
\usage{
neg_ll_composite(
  params,
  list_simu,
  df_lags,
  locations,
  quantile,
  list_excesses,
  latlon = FALSE,
  s0 = NA,
  t0 = NA,
  hmax = NA
)
}
\arguments{
\item{params}{Vector of variogram parameters (beta1, beta2, alpha1, alpha2).}

\item{list_simu}{A list of simulated data.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{locations}{The locations dataframe.}

\item{quantile}{The quantile value.}

\item{list_excesses}{A list of excesses dataframes.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is FALSE.}

\item{s0}{The starting location.}

\item{t0}{The starting time.}

\item{hmax}{The maximum spatial lag value.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate the negative log-likelihood for a list of simulations.
}
