% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll_composite_fixed_eta}
\alias{neg_ll_composite_fixed_eta}
\title{neg_ll_composite_fixed_eta function
 
Calculate the negative log-likelihood for a list of r-Pareto processes
with fixed advection parameters.}
\usage{
neg_ll_composite_fixed_eta(
  params,
  list_episodes,
  list_excesses,
  list_lags,
  wind_df = NA,
  hmax = NA,
  latlon = TRUE,
  distance = "euclidean",
  threshold = FALSE,
  rpar = TRUE,
  fixed_eta1 = NA,
  fixed_eta2 = NA
)
}
\arguments{
\item{params}{Vector of variogram parameters (beta1, beta2, alpha1
alpha2).}

\item{list_episodes}{A list of episode data.}

\item{list_excesses}{A list of excesses dataframes.}

\item{list_lags}{A list of dataframes with spatial and temporal lag values.}

\item{wind_df}{A dataframe with wind values. Default is NA.}

\item{hmax}{The maximum spatial lag value. Default is NA.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is TRUE.}

\item{distance}{The type of spatial norm, "euclidean" or "lalpha".
Default is "euclidean".}

\item{threshold}{A boolean value to indicate if the quantile variable is a
threshold value and not a uniform quantile. Default is FALSE.}

\item{rpar}{A boolean value to indicate if the data is considered as a
r-Pareto process. Default is TRUE.}

\item{fixed_eta1}{A value to fix eta1.}

\item{fixed_eta2}{A value to fix eta2.}
}
\value{
The negative log-likelihood value.
}
\description{
neg_ll_composite_fixed_eta function
 
Calculate the negative log-likelihood for a list of r-Pareto processes
with fixed advection parameters.
}
