% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll_composite_simu}
\alias{neg_ll_composite_simu}
\title{neg_ll_composite_simu function}
\usage{
neg_ll_composite_simu(
  params,
  list_simu,
  df_lags,
  quantile,
  list_excesses,
  hmax = NA,
  s0 = NA,
  t0 = NA,
  threshold = FALSE
)
}
\arguments{
\item{params}{Vector of variogram parameters (beta1, beta2, alpha1, alpha2).}

\item{list_simu}{A list of simulated data.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{quantile}{The quantile value.}

\item{list_excesses}{A list of excesses dataframes.}

\item{hmax}{The maximum spatial lag value.}

\item{s0}{The starting location.}

\item{t0}{The starting time.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate the negative log-likelihood for a list of simulations.
}
