% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll_composite_simu_par}
\alias{neg_ll_composite_simu_par}
\title{neg_ll_composite_simu_par function}
\usage{
neg_ll_composite_simu_par(
  beta1,
  beta2,
  alpha1,
  alpha2,
  adv1,
  adv2,
  list_simu,
  df_lags,
  quantile,
  list_excesses,
  hmax = NA,
  s0 = NA,
  t0 = NA,
  threshold = FALSE
)
}
\arguments{
\item{beta1}{The beta1 parameter.}

\item{beta2}{The beta2 parameter.}

\item{alpha1}{The alpha1 parameter.}

\item{alpha2}{The alpha2 parameter.}

\item{adv1}{The advection parameter 1.}

\item{adv2}{The advection parameter 2.}

\item{list_simu}{A list of simulated data.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{quantile}{The quantile value.}

\item{list_excesses}{A list of excesses dataframes.}

\item{hmax}{The maximum spatial lag value.}

\item{s0}{The starting location.}

\item{t0}{The starting time.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate the negative log-likelihood for a list of simulations.
Used for fixing parameters in the optimization process.
}
