% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll_par}
\alias{neg_ll_par}
\title{neg_ll_par function}
\usage{
neg_ll_par(
  beta1,
  beta2,
  alpha1,
  alpha2,
  adv1,
  adv2,
  data,
  df_lags,
  locations,
  quantile,
  excesses,
  latlon = FALSE,
  hmax = NA,
  s0 = NA,
  t0 = NA,
  threshold = FALSE,
  norm = "euclidean"
)
}
\arguments{
\item{beta1}{The beta1 parameter.}

\item{beta2}{The beta2 parameter.}

\item{alpha1}{The alpha1 parameter.}

\item{alpha2}{The alpha2 parameter.}

\item{adv1}{The advection parameter 1.}

\item{adv2}{The advection parameter 2.}

\item{data}{The data dataframe.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{locations}{The locations dataframe.}

\item{quantile}{The quantile value.}

\item{excesses}{The excesses dataframe with the number of excesses kij and
the number of possible excesses Tobs.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is FALSE.}

\item{hmax}{The maximum spatial lag value. Default is NA.}

\item{s0}{The conditioning location. Default is NA.}

\item{t0}{The conditioning time. Default is NA.}

\item{threshold}{A boolean value to indicate if the quantile variable is a
threshold value and not a uniform quantile. Default is FALSE.}

\item{pmarg}{The probability of marginal excesses. Default is NA.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate the negative log-likelihood for a given set of variogram
parameters by considering excess indicators following a binomial distribution
with a probability parameter equals to the theorical spatio-temporal chi
value. Used for fixing parameters in the optimization process.
}
