% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{norm_Lp}
\alias{norm_Lp}
\title{norm_Lp function}
\usage{
norm_Lp(x, y, p)
}
\arguments{
\item{x}{The first vector.}

\item{y}{The second vector.}

\item{p}{The exponent value.}
}
\value{
The Lp norm of the two vectors.
}
\description{
This function calculates the Lp norm of two vectors.
}
