% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{pEGPD_full}
\alias{pEGPD_full}
\title{pEGPD_full}
\usage{
pEGPD_full(x, p0, xi, sigma, kappa)
}
\arguments{
\item{x}{numeric vector of quantiles}

\item{p0}{probability of zero}

\item{xi}{shape parameter of the EGPD}

\item{sigma}{scale parameter of the EGPD}

\item{kappa}{second shape parameter of the EGPD}
}
\value{
numeric vector of CDF values
}
\description{
CDF of the entire distribution with a mass at zero and an EGPD for
positive values
}
