% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{plot_th_emp_chi}
\alias{plot_th_emp_chi}
\title{plot_th_emp_chi}
\usage{
plot_th_emp_chi(
  list_lags,
  list_excesses,
  list_adv,
  params_estimates,
  tau_min = 0,
  tau_fixed = 0,
  h_breaks = seq(0, 10, by = 1),
  latlon = FALSE,
  adv_transform = TRUE
)
}
\arguments{
\item{list_lags}{list of data frames with lag information for each episode}

\item{list_excesses}{list of data frames with excess information for each episode}

\item{list_adv}{data frame with advection parameters for each episode}

\item{params_estimates}{numeric vector of estimated model parameters}

\item{tau_min}{minimum tau value to consider}

\item{tau_fixed}{fixed tau value for specific plot}

\item{h_breaks}{breaks for h binning}

\item{latlon}{logical, whether coordinates are in lat/lon}

\item{adv_transform}{logical, whether to apply advection transformation}
}
\value{
invisible list with results and plots
}
\description{
Plot empirical vs theoretical chi values for multiple episodes
}
