% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{plot_transformation_gg}
\alias{plot_transformation_gg}
\title{plot_transformation_gg}
\usage{
plot_transformation_gg(Z, X, u, site_name, save_plot = FALSE, filename = NULL)
}
\arguments{
\item{Z}{matrix of latent variables (sites x time)}

\item{X}{matrix of observed variables (sites x time)}

\item{u}{threshold in the latent space}

\item{site_name}{name of the site to plot}

\item{save_plot}{logical, whether to save the plot}

\item{filename}{filename to save the plot if save_plot is TRUE}
}
\value{
ggplot object
}
\description{
Plot the transformation from latent Z to observed X for a given site
}
