% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{process_simulation}
\alias{process_simulation}
\title{process_simulation function}
\usage{
process_simulation(
  i,
  m,
  list_simu,
  u,
  list_lags,
  list_excesses,
  init_params,
  hmax = NA,
  wind_df = NA,
  distance = "euclidean",
  normalize = FALSE,
  fixed_eta1 = NA,
  fixed_eta2 = NA
)
}
\arguments{
\item{i}{The index of the simulation.}

\item{m}{The number of simulations to process.}

\item{list_simu}{The list of simulations.}

\item{u}{The quantile threshold.}

\item{list_lags}{The list of lags.}

\item{list_excesses}{The list of excesses.}

\item{init_params}{The initial parameters for optimization.}

\item{hmax}{The maximum spatial lag value. Default is NA.}

\item{wind_df}{The wind dataframe. Default is NA.}

\item{distance}{The type of spatial norm, "euclidean" or "lalpha". 
Default is "euclidean".}

\item{hessian}{A boolean value to indicate if the Hessian matrix should be
computed. Default is FALSE.}
}
\value{
The optimized parameters.
}
\description{
This function processes the simulation data and optimizes the parameters
using the negative log-likelihood function.
}
