% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd.R
\name{process_site}
\alias{process_site}
\title{process_site function
This function fits the Extreme Generalized Pareto Distribution (EGPD)
to the given data for a specific site and generates a QQ plot.}
\usage{
process_site(y, site_name, save_path = "", R = 1000, forced_cens = NULL)
}
\arguments{
\item{y}{The data to be fitted.}

\item{site_name}{The name of the site}

\item{save_path}{The path to save the QQ plot. Default is "".}

\item{R}{The number of bootstrap samples for confidence intervals. Default is 1000}

\item{forced_cens}{The forced censoring value. Default is NULL.}
}
\value{
A data frame containing the fitted parameters and best censore for the site.
}
\description{
process_site function
This function fits the Extreme Generalized Pareto Distribution (EGPD)
to the given data for a specific site and generates a QQ plot.
}
