% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{qEGPD_full}
\alias{qEGPD_full}
\title{qEGPD_full}
\usage{
qEGPD_full(v, p0, xi, sigma, kappa)
}
\arguments{
\item{v}{numeric vector of probabilities}

\item{p0}{probability of zero}

\item{xi}{shape parameter of the EGPD}

\item{sigma}{scale parameter of the EGPD}

\item{kappa}{second shape parameter of the EGPD}
}
\value{
numeric vector of quantiles
}
\description{
Quantile function of the entire distribution with a mass at zero and an EGPD for
positive values
}
