% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile_matrix}
\alias{quantile_matrix}
\title{This function calculates the quantile matrix for a given quantile q, using
the data_rain dataset.
The count_min parameter specifies the minimum number of observations
required to calculate a quantile.
The step parameter determines the resolution of the quantile matrix.}
\usage{
quantile_matrix(
  q,
  data_rain,
  count_min = 80,
  step = 0.005,
  zeros = TRUE,
  qlim = TRUE
)
}
\arguments{
\item{q}{The quantile value.}

\item{data_rain}{The dataset containing the rainfall data.}

\item{count_min}{The minimum number of observations required to calculate a
quantile. Default is 80.}

\item{step}{The resolution of the quantile matrix. Default is 0.005.}

\item{zeros}{A logical value indicating whether to include zero values in the
dataset. Default is TRUE.}

\item{qlim}{A logical value indicating a quantile limit, more precised some
times. Default is TRUE.}
}
\value{
A matrix containing the quantile values for each pair of variables
in the dataset.
}
\description{
This function calculates the quantile matrix for a given quantile q, using
the data_rain dataset.
The count_min parameter specifies the minimum number of observations
required to calculate a quantile.
The step parameter determines the resolution of the quantile matrix.
}
