% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{reshape_distances}
\alias{reshape_distances}
\title{This function reshapes a distance matrix into a long dataframe.}
\usage{
reshape_distances(dist_mat)
}
\arguments{
\item{dist_mat}{The input distance matrix.}
}
\value{
A dataframe containing the reshaped distances.
}
\description{
This function reshapes a distance matrix into a long dataframe.
}
