% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{save_simulations}
\alias{save_simulations}
\title{save_simulations function}
\usage{
save_simulations(BR, ngrid, n.BR, folder, file = "rainBR", forcedind = NA)
}
\arguments{
\item{BR}{The BR simulationsas array.}

\item{ngrid}{The number of grid points.}

\item{n.BR}{The number of BR simulations.}

\item{folder}{The folder path.}

\item{file}{The filename without extension, default is "rainBR".}

\item{forcedind}{The index of the simulation to save, default is NA.}
}
\value{
None
}
\description{
This function transforms the Brown-Resnick simulations into dataframes and
save them into CSV files rain_BR_i.csv.
}
