% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_tablatex}
\alias{save_tablatex}
\title{Save a LaTeX table to a file}
\usage{
save_tablatex(
  table_latex,
  filename = "",
  pval = FALSE,
  tab_folder = "/tables/"
)
}
\arguments{
\item{table_latex}{The LaTeX code representing the table.}

\item{filename}{The name of the file to save the table to. If not provided,
the table will not be saved to a file.}

\item{pval}{A logical value indicating whether to include p-values
in the table. Default is FALSE.}

\item{tab_folder}{The folder where the table will be saved.
Default is '/tables/'.}
}
\value{
None
}
\description{
This function saves a LaTeX table to a file.
}
