% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{sd_vario}
\alias{sd_vario}
\title{Calculate the standard deviation of a variogram}
\usage{
sd_vario(x, vario, sd_c, sd_alpha)
}
\arguments{
\item{x}{The input data.}

\item{vario}{The variogram function.}

\item{sd_c}{The constant term in the standard deviation formula.}

\item{sd_alpha}{The alpha term in the standard deviation formula.}
}
\value{
The standard deviation of the variogram.
}
\description{
This function calculates the standard deviation of a variogram based on
the given parameters.
}
\examples{
x <- c(1, 2, 3, 4, 5)
vario <- function(h) h^2
sd_c <- 1
sd_alpha <- 0.5
sd_vario(x, vario, sd_c, sd_alpha)

}
