% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{select_max_extreme_episodes}
\alias{select_max_extreme_episodes}
\title{select_max_extreme_episodes function}
\usage{
select_max_extreme_episodes(
  sites_coords,
  data,
  quantile,
  min_spatial_dist,
  delta,
  n_max_episodes = 10000,
  time_ext = 0,
  latlon = TRUE
)
}
\arguments{
\item{sites_coords}{The coordinates of the sites.}

\item{data}{The rainfall dataframe.}

\item{quantile}{The quantile value.}

\item{min_spatial_dist}{The minimum spatial distance between two episodes.}

\item{delta}{The temporal window size.}

\item{n_max_episodes}{The maximum number of episodes to select. Default is
10000.}

\item{time_ext}{The temporal window extension. Default is 0.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is TRUE.}
}
\value{
A list of selected points (s0, t0) and the corresponding threshold 
        value u_s0 for each episode.
}
\description{
This function selects extreme episodes based on the given data and a 
quantile. A spatio-temporal specific neighborhood is considered to avoid
selecting nearby episodes. Each episode is selected according to a 
r-Pareto process with r(X) = max(X(s, t)). Each episode will have a size of
2 * delta.
}
