% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_BR_parallel}
\alias{sim_BR_parallel}
\title{sim_BR_parallel function}
\usage{
sim_BR_parallel(params, n.BR, spa, temp)
}
\arguments{
\item{params}{The parameters for the simulation.}

\item{n.BR}{The number of BR simulations to perform.}

\item{spa}{The spatial data for the simulation.}

\item{temp}{The temporal data for the simulation.}
}
\value{
The results of the parallel simulations.
}
\description{
This function performs parallel simulations using the BR method.
}
