% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{sim_episode_coords}
\alias{sim_episode_coords}
\title{sim_episode_coords}
\usage{
sim_episode_coords(
  params_vario,
  params_margins,
  coords,
  times,
  adv,
  t0,
  s0,
  u,
  u_emp,
  plot_debug = FALSE,
  filename = NULL
)
}
\arguments{
\item{params_vario}{list of variogram parameters}

\item{params_margins}{list of marginal parameters}

\item{coords}{matrix of coordinates (sites x 2)}

\item{times}{vector of times}

\item{adv}{vector of advection (vx, vy)}

\item{t0}{time index of the event origin}

\item{s0}{site name of the event origin}

\item{u}{threshold in the latent space}

\item{u_emp}{empirical threshold in the observed space}

\item{plot_debug}{logical, whether to plot debug plots}

\item{filename}{filename prefix to save debug plots if plot_debug is TRUE}
}
\value{
list with latent Z, observed X, and latent threshold u_latent
}
\description{
Simulate a single episode over given coordinates and times
with r-Pareto process and transform to observed scale
}
