% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{sim_episode_grid}
\alias{sim_episode_grid}
\title{sim_episode_grid}
\usage{
sim_episode_grid(
  params_vario,
  params_margins_common,
  coords,
  times,
  adv,
  t0,
  s0_pixel_id,
  u,
  u_emp,
  plot_debug = FALSE,
  filename = NULL
)
}
\arguments{
\item{params_vario}{list of variogram parameters}

\item{params_margins_common}{list of common marginal parameters}

\item{coords}{matrix of coordinates (sites x 2)}

\item{times}{vector of times}

\item{adv}{vector of advection (vx, vy)}

\item{t0}{time index of the event origin}

\item{s0_pixel_id}{site name of the event origin}

\item{u}{threshold in the latent space}

\item{u_emp}{empirical threshold in the observed space}

\item{plot_debug}{logical, whether to plot debug plots}

\item{filename}{filename prefix to save debug plots if plot_debug is TRUE}
}
\value{
matrix of observed X values (sites x time)
}
\description{
Simulate a single episode over given regular grid coordinates and times
with r-Pareto process and transform to observed scale
}
