% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_rpareto}
\alias{sim_rpareto}
\title{sim_rpareto function (extended)}
\usage{
sim_rpareto(
  beta1,
  beta2,
  alpha1,
  alpha2,
  x,
  y,
  t,
  adv = c(0, 0),
  t0 = 0,
  nres = 1,
  random_s0 = FALSE,
  s0 = c(1, 1),
  s0_radius = Inf,
  distance = "euclidean",
  seed = NULL
)
}
\arguments{
\item{beta1, }{beta2 Variogram scale parameters for space and time.}

\item{alpha1, }{alpha2 Variogram smoothness parameters for space and time.}

\item{x, }{y, t Vectors representing spatial (x, y) and temporal (t) grids.}

\item{adv}{Advection vector (default = c(0, 0)).}

\item{t0}{Time point at which the process is conditioned.}

\item{nres}{Number of simulations to perform.}

\item{random_s0}{If TRUE, selects conditioning point(s) randomly within radius.}

\item{s0}{Conditioning spatial location(s). 
Either a vector c(x,y) for one point, or a matrix/data.frame with 2 columns for multiple points.}

\item{s0_radius}{Radius used if random_s0 is TRUE.}

\item{distance}{Type of distance metric to use ("euclidean" or "lalpha").}
}
\value{
A list with simulated field Z (5D array: x × y × t × nres × n_s0)
        and list of conditioning points s0_used.
}
\description{
This function simulates a spatio-temporal r-Pareto process using either 
a fractional Brownian motion model.
It supports conditioning on one or multiple spatial points s0.
}
