% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_rpareto_aniso}
\alias{sim_rpareto_aniso}
\title{Simulate the spatio-temporal r-Pareto process with anisotropy
and advection.}
\usage{
sim_rpareto_aniso(
  beta1,
  beta2,
  alpha1,
  alpha2,
  x,
  y,
  t,
  adv = c(0, 0),
  s0 = c(1, 1),
  t0 = 1,
  nres = 1
)
}
\arguments{
\item{beta1}{The value of beta1.}

\item{beta2}{The value of beta2.}

\item{alpha1}{The value of alpha1.}

\item{alpha2}{The value of alpha2.}

\item{x}{Vector for the first dimension (spatial x in our case).}

\item{y}{Vector for the second dimension (spatial y in our case)}

\item{t}{Vector for the third dimension (time in our case).}

\item{adv}{The advection coordinates vector. Default is c(0, 0).}
}
\value{
The result of the simulation.
}
\description{
Simulate the spatio-temporal r-Pareto process with anisotropy
and advection.
}
