% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_rpareto_coords}
\alias{sim_rpareto_coords}
\title{sim_rpareto_coords function
This function simulates a spatio-temporal r-Pareto process using a
fractional Brownian motion model on given coordinates grid, regular or not.}
\usage{
sim_rpareto_coords(
  coords,
  times,
  beta1,
  beta2,
  alpha1,
  alpha2,
  adv = c(0, 0),
  threshold = 1,
  s0_index = 1,
  t0_index = 1,
  seed = NULL
)
}
\arguments{
\item{coords}{Data frame or matrix with columns 'Longitude' and 'Latitude'.}

\item{times}{Vector of time points.}

\item{beta1, }{beta2 Variogram scale parameters for space and time.}

\item{alpha1, }{alpha2 Variogram smoothness parameters for space and time.}

\item{adv}{Advection vector (default = c(0, 0)).}

\item{threshold}{Threshold for the r-Pareto process (default = 1).}

\item{s0_index}{Index of the spatial conditioning point.}

\item{t0_index}{Index of the temporal conditioning point.}

\item{seed}{Optional random seed for reproducibility.}
}
\value{
A list with simulated field Z, Gaussian process W, and variogram gamma0.
}
\description{
sim_rpareto_coords function
This function simulates a spatio-temporal r-Pareto process using a
fractional Brownian motion model on given coordinates grid, regular or not.
}
