% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{simulate_excess_ind}
\alias{simulate_excess_ind}
\title{simulate_excess_ind function}
\usage{
simulate_excess_ind(Tmax, chi_h_t, p_marg)
}
\arguments{
\item{Tmax}{The maximum time observation.}

\item{chi_h_t}{The spatio-temporal extremogram value at a given spatial lag h
and a given temporal lag t.}

\item{p_marg}{The probability of marginal excesses.}
}
\value{
The simulated individual excess.
}
\description{
Simulate individual excess following a binomial distribution with a given
probability parameter equals to the chi value.
}
