% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{simulate_many_episodes}
\alias{simulate_many_episodes}
\title{simulate_many_episodes}
\usage{
simulate_many_episodes(
  N,
  u,
  u_emp,
  params_vario,
  params_margins,
  coords,
  times,
  adv,
  t0,
  s0,
  plot_debug = FALSE
)
}
\arguments{
\item{N}{number of episodes to simulate}

\item{u}{latent threshold}

\item{u_emp}{empirical threshold}

\item{params_vario}{variogram parameters}

\item{params_margins}{marginal parameters}

\item{coords}{coordinates of sites}

\item{times}{time points}

\item{adv}{advection parameters}

\item{t0}{initial time index}

\item{s0}{initial site}

\item{plot_debug}{whether to plot debug information}
}
\value{
list of simulated episodes
}
\description{
Simulate multiple episodes of rainfall data
}
