% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{spatial_chi_pair}
\alias{spatial_chi_pair}
\title{Calculate the spatial chi statistic}
\usage{
spatial_chi_pair(rain_pair, q, zeros = TRUE)
}
\arguments{
\item{rain_pair}{The rainfall data for a pair of sites.}

\item{q}{The quantile value.}

\item{zeros}{Logical indicating whether to include zero values in the
calculation.}
}
\value{
The spatial chi statistic.
}
\description{
This function calculates the spatial chi statistic for a given dataset
and a quantile value.
}
