% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{spatial_mean_lags}
\alias{spatial_mean_lags}
\title{Calculate the spatial mean lags}
\usage{
spatial_mean_lags(radius, mid = FALSE)
}
\arguments{
\item{radius}{The radius used to calculate the spatial mean lags.}

\item{mid}{Logical value indicating whether to return the midpoint of the 
lags. Default is FALSE.}
}
\value{
A vector of spatial mean lags or the midpoint of the lags if 'mid' 
        is TRUE.
}
\description{
This function calculates the spatial mean lags based on a given radius.
The function takes an optional argument 'mid' which determines whether
the midpoint of the lags should be returned instead of the full lags.
}
\examples{
spatial_mean_lags(5)
spatial_mean_lags(10, mid = TRUE)

}
