% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg.R
\name{sum_over_time_by_site}
\alias{sum_over_time_by_site}
\title{sum_over_time_by_site}
\usage{
sum_over_time_by_site(ep_mat, sites_ref)
}
\arguments{
\item{ep_mat}{matrix of episode values (time x site) or (site x time)}

\item{sites_ref}{vector of site names}
}
\value{
named numeric vector of summed values by site
}
\description{
Sum values over time for each site in an episode
}
