% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{temporal_chi}
\alias{temporal_chi}
\title{Calculate the temporal chi statistic}
\usage{
temporal_chi(data_rain, tmax, quantile, zeros = TRUE, mean = TRUE)
}
\arguments{
\item{data_rain}{The rainfall data.}

\item{tmax}{The maximum threshold value.}

\item{quantile}{The quantile value.}

\item{zeros}{Logical indicating whether to include zero values in the
calculation.}

\item{mean}{Logical indicating whether to return the mean value.}
}
\value{
The temporal chi statistic.
}
\description{
This function calculates the temporal extremogram for a given dataset.
}
\examples{
data <- c(1, 2, 3, 4, 5)
temporal_chi(data, 3, 0.5)

}
