% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{theoretical_chi_aniso}
\alias{theoretical_chi_aniso}
\title{theoretical_chi_aniso function}
\usage{
theoretical_chi_aniso(params, df_lags, latlon = FALSE, wind_vect = NA)
}
\arguments{
\item{params}{A vector of parameters.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is FALSE.}

\item{wind_vect}{A vector of wind values. Default is NA.}
}
\value{
The theoretical spatio-temporal extremogram.
}
\description{
This function calculates the theoretical spatio-temporal extremogram
based on the given parameters.
}
