% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{theorical_chi}
\alias{theorical_chi}
\title{Compute the theoretical chi dataframe.}
\usage{
theorical_chi(params, df_lags)
}
\arguments{
\item{params}{A vector of parameters.}

\item{df_lags}{A dataframe with spatial lag values.}

\item{tau}{A vector of temporal lag values.}
}
\value{
The theoretical chi matrix.
}
\description{
This function calculates the theoretical chi dataframe. based on the given
parameters.
}
