% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R, R/simulations.R
\name{vario}
\alias{vario}
\title{Calculate the spatial or temporal variogram}
\usage{
vario(h, tau, beta1, beta2, alpha1, alpha2)

vario(h, tau, beta1, beta2, alpha1, alpha2)
}
\arguments{
\item{h}{The spatial lags.}

\item{tau}{The temporal lags.}

\item{beta1}{The value of beta1.}

\item{beta2}{The value of beta2.}

\item{alpha1}{The value of alpha1.}

\item{alpha2}{The value of alpha2.}

\item{x}{The corresponding lags.}

\item{c}{The c variogram parameter. (log(beta))}

\item{alpha}{The alpha variogram parameter.}
}
\value{
The calculated spatial or temporal variogram.

The variogram value.
}
\description{
This function calculates the spatial or temporal variogram based on the given
WLSE estimated parameters c and alpha. (c = log(beta))

This function computes the variogram value for a given distance matrix and 
variogram parameters.
}
\examples{
x <- c(1, 2, 3, 4, 5)
c <- 1
alpha <- 0.5
vario <- vario(x, c, alpha)
print(vario)
# Output: [1] 2.718282 7.389056 20.085537 54.598150 148.413159
}
