% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{vario_spatemp}
\alias{vario_spatemp}
\title{Calculate the theorical spatio-temporal variogram}
\usage{
vario_spatemp(chi)
}
\arguments{
\item{chi}{The extremogram.}
}
\value{
The theorical spatio-temporal variogram.
}
\description{
This function calculates the theorical spatio-temporal variogram based on the
given extremogram.
}
