% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wlse_tablatex}
\alias{wlse_tablatex}
\title{wlse_tablatex function}
\usage{
wlse_tablatex(summary_df, ind, filename = "")
}
\arguments{
\item{summary_df}{The summary dataframe containing the data to be converted
into a table.}

\item{ind}{The index of the summary dataframe to be used for generating the
table.}

\item{filename}{The name of the file to save the generated LaTeX table.
If not provided, the table will be printed to the console.}
}
\value{
The generated LaTeX table.
}
\description{
This function generates a LaTeX table from a summary dataframe of WLSE
results.
}
