% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{chispatemp_dt}
\alias{chispatemp_dt}
\title{chispatemp_dt function}
\usage{
chispatemp_dt(lagt, lags, rad_mat, data_rain, quant_mat = NA, q_fixed = NA)
}
\arguments{
\item{lagt}{The temporal lag.}

\item{lags}{The spatial lags.}

\item{rad_mat}{The matrix of radii.}

\item{data_rain}{The rainfall data.}

\item{quant_mat}{The matrix of quantiles (optional).}

\item{q_fixed}{The fixed quantile value (optional).}
}
\value{
The result of the chi spatial-temporal dependence.
}
\description{
This function calculates the chi spatial-temporal dependence
}
