% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{compute_wind_episode}
\alias{compute_wind_episode}
\title{compute_wind_episode function}
\usage{
compute_wind_episode(episode, s0, u, wind_df, speed_time)
}
\arguments{
\item{episode}{The extreme episode dataframe.}

\item{s0}{The conditioning location name.}

\item{u}{The quantile threshold.}

\item{wind_df}{The wind dataframe.}

\item{speed_time}{The time index for wind speed.}
}
\value{
The wind episode dataframe.
}
\description{
Compute the wind speed and direction for a given extreme episode.
}
