% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_simu_gif}
\alias{create_simu_gif}
\title{Create a GIF from a simulation}
\usage{
create_simu_gif(simulation_data, params, type = "rpar", forcedtemp = NA)
}
\arguments{
\item{simulation_data}{The data from the simulation.}

\item{params}{The parameters used for the simulation.}

\item{type}{The type of simulation.}

\item{forcedtemp}{The number of time steps to include in the GIF.}
}
\value{
None, but saves the GIF to a file.
}
\description{
This function creates a GIF from a simulation.
}
