% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{empirical_excesses}
\alias{empirical_excesses}
\title{empirical_excesses function}
\usage{
empirical_excesses(data_rain, quantile, h_vect)
}
\arguments{
\item{data_rain}{The rainfall data.}

\item{quantile}{The quantile value.}

\item{h_vect}{The spatial lag vectors and distances dataframe.}

\item{tau}{The temporal lag vector.}

\item{nmin}{The minimum number of observations to consider. Default is 5.}
}
\value{
A list with n_vect the number of excesses and N_vect the number of
possible excesses ie the number of observations for each pair of sites.
}
\description{
This function calculates the empirical excesses based on indicators above a
quantile threshold.
}
