% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{empirical_excesses}
\alias{empirical_excesses}
\title{empirical_excesses function}
\usage{
empirical_excesses(data_rain, quantile, tau, h_vect, df_dist)
}
\arguments{
\item{data_rain}{The rainfall data.}

\item{quantile}{The quantile value.}

\item{tau}{The temporal lag vector.}

\item{h_vect}{The spatial lag vector.}

\item{df_dist}{The distances inside a long dataframe.}
}
\value{
A list with n_vect the number of excesses and N_vect the number of
possible excesses ie the number of observations for each pair of sites.
}
\description{
This function calculates the empirical excesses based on indicators above a
quantile threshold.
}
\examples{
empirical_excesses(data_rain, 0.95, 1:10, c(0.1, 0.2, 0.3), df_dist)
}
