% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{empirical_excesses_rpar}
\alias{empirical_excesses_rpar}
\title{empirical_excesses_rpar function}
\usage{
empirical_excesses_rpar(
  data_rain,
  quantile,
  df_lags,
  threshold = FALSE,
  t0 = 1
)
}
\arguments{
\item{data_rain}{The rainfall data.}

\item{quantile}{The quantile value.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{threshold}{A boolean value to indicate if the quantile variable is a
threshold value and not a uniform quantile. Default is FALSE.}

\item{t0}{The starting time. Default is 1.}
}
\value{
The empirical excesses dataframe with the number of excesses kij and
the number of possible excesses Tobs, with the lag values.
}
\description{
This function calculates the empirical excesses based on indicators above a
quantile threshold for the r-Pareto process.
}
