% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{eta}
\alias{eta}
\title{Transforms the extremogram using the eta transformation.}
\usage{
eta(chi)
}
\arguments{
\item{chi}{The input extremogram to be transformed.}
}
\value{
The transformed vector.
}
\description{
The eta function transforms the input vector by applying a logarithmic
transformation using the quantile function of the standard normal
distribution.
}
