% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{evaluate_optim}
\alias{evaluate_optim}
\title{evaluate_optim function}
\usage{
evaluate_optim(
  list_simu,
  quantile,
  true_param,
  df_lags,
  locations,
  parscale = c(1, 1, 1, 1),
  latlon = FALSE
)
}
\arguments{
\item{list_simu}{A list of simulated data}

\item{quantile}{The quantile value}

\item{true_param}{The true variogram parameter (beta1, beta2, alpha1, alpha2)}

\item{locations}{The locations dataframe}

\item{parscale}{The scaling parameter for the optimization process. Default
is c(1, 1, 1, 1).}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is FALSE.}

\item{tau_vect}{The temporal lag vector}

\item{hmax}{The maximum spatial lag value. Default is sqrt(17).}

\item{nmin}{The minimum number of observations to consider}
}
\value{
The result of the optimization process as a dataframe.
}
\description{
This function evaluates the optimization process on simulated data (for
example, following a Brown-Resnick process) using the negative
log-likelihood function.
}
