% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_estimate}
\alias{format_estimate}
\title{Formats p-values with asterisks}
\usage{
format_estimate(estimate, p_value)
}
\arguments{
\item{estimates}{A numeric vector of coefficient estimates.}

\item{p_values}{A numeric vector of p-values corresponding to the
coefficient estimates.}
}
\value{
A character vector of formatted coefficient estimates
         with asterisks indicating significance.
}
\description{
This function formats coefficient estimates by adding asterisks
to indicate significance.
}
\examples{
estimates <- c(0.05, 0.01, 0.001)
p_values <- c(0.1, 0.05, 0.001)
format_coefficients(estimates, p_values, 0.05)
# Output: "0.05", "0.01*", "0.001**"

}
