% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_variogram_plots}
\alias{generate_variogram_plots}
\title{Generate variogram plots
This function generates variogram plots for multiple tau values.}
\usage{
generate_variogram_plots(result, df_lags, true_param, tau_values, chi = FALSE)
}
\arguments{
\item{result}{The result of the variogram estimation.}

\item{df_lags}{The dataframe of lags.}

\item{true_param}{The true parameters used to generate the data.}

\item{tau_values}{The values of tau to generate plots for.}

\item{chi}{A logical value indicating whether to plot the chi values.}
}
\value{
None, but generates plots.
}
\description{
Generate variogram plots
This function generates variogram plots for multiple tau values.
}
