% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{get_chiq}
\alias{get_chiq}
\title{This function calculates the chi extremogram for a given dataset
and a quantile value.}
\usage{
get_chiq(data, quantile)
}
\arguments{
\item{data}{The data set.}

\item{quantile}{The quantile value used in the calculation.}
}
\value{
The chi value.
}
\description{
This function calculates the chi extremogram for a given dataset
and a quantile value.
}
\examples{
data1 <- c(1, 2, 3, 4, 5)
data2 <- c(2, 3, 4, 5, 6)
data <- cbind(data1, data2)
quantile <- 0.95
get_chiq(data, quantile)

}
