% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{get_chiq}
\alias{get_chiq}
\title{This function calculates the chi extremogram for a given dataset
and a quantile value.}
\usage{
get_chiq(data, quantile)
}
\arguments{
\item{data}{The data set.}

\item{quantile}{The quantile value.}
}
\value{
The chi value.
}
\description{
This function calculates the chi extremogram for a given dataset
and a quantile value.
}
