% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_conditional_lag_vectors}
\alias{get_conditional_lag_vectors}
\title{get_conditional_lag_vectors function}
\usage{
get_conditional_lag_vectors(
  df_coords,
  params,
  temp,
  s0 = c(1, 1),
  t0 = 1,
  tau_max = 10
)
}
\arguments{
\item{df_coords}{A dataframe containing the coordinates of the points.}

\item{params}{A vector of parameters.}

\item{temp}{The temporal sequence.}

\item{s0}{Conditional spatial point coordinates.}

\item{t0}{Conditional temporal point.}

\item{tau_max}{The maximum temporal lag. Default is 10.}
}
\value{
A dataframe containing the lag vectors.
}
\description{
This function calculates the lag vectors between pairs of points.
}
