% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_dist_mat}
\alias{get_dist_mat}
\title{Compute the distance matrix between locations}
\usage{
get_dist_mat(locations, dmax = NA, latlon = TRUE)
}
\arguments{
\item{locations}{A matrix or data frame containing the coordinates of the
locations.}

\item{dmax}{The maximum distance threshold if specified.}

\item{latlon}{Logical indicating whether the coordinates are in latitude and
longitude format.}
}
\value{
A distance matrix where each element represents the distance between
two locations.
}
\description{
This function computes the distance matrix between a set of locations.
}
