% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{get_estimate_variotemp}
\alias{get_estimate_variotemp}
\title{Calculate the estimate of variotemp}
\usage{
get_estimate_variotemp(chitemp, tmax, npoints, weights, summary = FALSE)
}
\arguments{
\item{chitemp}{The chitemp parameter.}

\item{tmax}{The tmax parameter.}

\item{npoints}{The number of points parameter.}

\item{weights}{The weights parameter, "residuals", "exp", or "none".}

\item{summary}{Logical value indicating whether to return the summary of the
model. Default is FALSE.}
}
\value{
The estimate of variotemp.
}
\description{
This function calculates the estimate of variotemp based on the given
parameters.
}
