% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_extreme_episodes}
\alias{get_extreme_episodes}
\title{get_extreme_episodes function}
\usage{
get_extreme_episodes(selected_points, data, delta, beta = 0, unif = FALSE)
}
\arguments{
\item{selected_points}{The selected points (s0, t0).}

\item{data}{The rainfall data.}

\item{delta}{The temporal window size.}

\item{beta}{The temporal window extension. Default is 0.}

\item{unif}{A boolean value to indicate if we want to get the uniformized
data episodes. Default is FALSE.}
}
\value{
The list of extreme episodes of size 2 * delta each.
}
\description{
This function extracts the extreme episodes based on the selected points
and the data according to r-Pareto process definition with r(X) = X(s0, t0).
}
