% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_h_vect}
\alias{get_h_vect}
\title{get_h_vect function}
\usage{
get_h_vect(dist, hmax = NA, intervals = FALSE)
}
\arguments{
\item{hmax}{The maximum spacial lag value. Default is NA.}

\item{intervals}{A logical value indicating whether we work with intervals.}

\item{df_dist}{The distance dataframe or matrix for intervals.}
}
\value{
The spatial distance lags vector or list of lag vectors.
}
\description{
This function calculates the spatial lag vector based on the given distance
dataframe and maximum spacial lag value.
}
