% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_h_vect}
\alias{get_h_vect}
\title{get_h_vect function}
\usage{
get_h_vect(df_dist, hmax)
}
\arguments{
\item{df_dist}{The distance dataframe.}

\item{hmax}{The maximum spacial lag value.}
}
\value{
The h vector.
}
\description{
This function calculates the spatial lag vector based on the given distance 
dataframe and maximum spacial lag value.
}
