% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_lag_vectors}
\alias{get_lag_vectors}
\title{get_lag_vectors function}
\usage{
get_lag_vectors(df_coords, tau_vect, latlon = FALSE)
}
\arguments{
\item{df_coords}{A dataframe containing the coordinates of the points.}

\item{tau_vect}{A vector of temporal lags.}

\item{latlon}{Logical indicating whether the coordinates are in latitude and
longitude format. Default is FALSE.}
}
\value{
A dataframe containing the lag vectors.
}
\description{
This function calculates the lag vectors between pairs of points.
}
