% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_lag_vectors}
\alias{get_lag_vectors}
\title{get_lag_vectors function}
\usage{
get_lag_vectors(
  df_coords,
  params,
  tau_vect = 0:10,
  hmax = NA,
  norm = "euclidean",
  index_s0 = 0
)
}
\arguments{
\item{df_coords}{A dataframe containing the coordinates of the points.}

\item{params}{A vector of parameters.}

\item{tau_vect}{A vector of temporal lags. Default is 0:10.}

\item{hmax}{The maximum distance threshold. Default is NA.}

\item{norm}{The norm to use. Can be "euclidean" or "Lp". Default
is "euclidean".}
}
\value{
A dataframe containing the lag vectors.
}
\description{
This function calculates the lag vectors between pairs of points.
}
