% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_marginal_excess}
\alias{get_marginal_excess}
\title{get_marginal_excess function}
\usage{
get_marginal_excess(
  data_rain,
  quantile,
  ind_s0 = NA,
  t0 = NA,
  threshold = FALSE
)
}
\arguments{
\item{data_rain}{The rainfall data.}

\item{quantile}{The quantile value.}
}
\value{
The number of marginal excesses.
}
\description{
This function calculates the number of marginal excesses above a given
quantile threshold.
}
