% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll}
\alias{neg_ll}
\title{neg_ll function}
\usage{
neg_ll(
  params,
  data,
  df_lags,
  locations,
  quantile,
  excesses,
  latlon = FALSE,
  simu_exp = FALSE
)
}
\arguments{
\item{params}{Vector of variogram parameters (beta1, beta2, alpha1, alpha2).}

\item{data}{The data dataframe.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{locations}{The locations dataframe.}

\item{quantile}{The quantile value.}

\item{excesses}{The excesses dataframe with the number of excesses kij and
the number of possible excesses Tobs.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is FALSE.}

\item{simu_exp}{A boolean value to indicate if the data is simulated with
an exponential distribution.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate the negative log-likelihood for a given set of variogram
parameters by considering excess indicators following a binomial distribution
with a probability parameter equals to the theorical spatio-temporal chi
value.
}
