% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll}
\alias{neg_ll}
\title{neg_ll function}
\usage{
neg_ll(
  params,
  df_lags,
  excesses,
  wind_vect = NA,
  hmax = NA,
  rpar = TRUE,
  threshold = FALSE,
  latlon = FALSE,
  quantile = NA,
  data = NA
)
}
\arguments{
\item{params}{Vector of variogram parameters (beta1, beta2, alpha1, alpha2).}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{excesses}{The excesses dataframe with the number of excesses kij and
the number of possible excesses Tobs.}

\item{wind_vect}{A vector of wind values. Default is NA.}

\item{hmax}{The maximum spatial lag value. Default is NA.}

\item{rpar}{A boolean value to indicate if the data is considered as a
r-Pareto process. Default is TRUE.}

\item{threshold}{A boolean value to indicate if the quantile variable is a
threshold value and not a uniform quantile. Default is FALSE.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is FALSE.}

\item{quantile}{The quantile value. Default is NA.}

\item{data}{The data dataframe (brown-resnick case). Default is NA.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate the negative log-likelihood for a given set of variogram
parameters by considering excess indicators following a binomial distribution
with a probability parameter equals to the theoretical spatio-temporal chi
value.
}
