% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll_composite}
\alias{neg_ll_composite}
\title{neg_ll_composite function}
\usage{
neg_ll_composite(
  params,
  list_lags,
  list_excesses,
  wind_df = NA,
  hmax = NA,
  latlon = TRUE,
  directional = TRUE
)
}
\arguments{
\item{params}{Vector of variogram parameters (beta1, beta2, alpha1, alpha2
eta1, eta2).}

\item{list_lags}{A list of dataframes with spatial and temporal lag values.}

\item{list_excesses}{A list of excesses dataframes.}

\item{wind_df}{A dataframe with wind values. Default is NA.}

\item{hmax}{The maximum spatial lag value. Default is NA.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is TRUE.}

\item{directional}{A boolean value to indicate if the variogram is
directional. Default is TRUE.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate the negative log-likelihood for a list of r-Pareto processes
with wind covariates if wanted.
}
