% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{neg_ll_composite_simu}
\alias{neg_ll_composite_simu}
\title{neg_ll_composite_simu function}
\usage{
neg_ll_composite_simu(
  params,
  list_simu,
  df_lags,
  quantile,
  list_excesses,
  hmax = NA,
  threshold = FALSE,
  directional = FALSE,
  rpar = TRUE
)
}
\arguments{
\item{params}{Vector of variogram parameters (beta1, beta2, alpha1, alpha2).}

\item{list_simu}{A list of simulated data.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{quantile}{The quantile value.}

\item{list_excesses}{A list of excesses dataframes.}

\item{hmax}{The maximum spatial lag value.}

\item{threshold}{A boolean value to indicate if the quantile variable is a
threshold value and not a uniform quantile. Default is FALSE.}

\item{directional}{A boolean value to indicate if the variogram is
directional. Default is TRUE.}

\item{rpar}{A boolean value to indicate if the data is considered as an
r-Pareto process. Default is TRUE.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate the negative log-likelihood for a list of simulations.
}
