% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{process_simulation}
\alias{process_simulation}
\title{process_simulation function}
\usage{
process_simulation(i, M, m, list_simuM, u, df_lags, t0, true_param, hmax = NA)
}
\arguments{
\item{i}{The index of the simulation.}

\item{M}{The number of simulations.}

\item{m}{The number of simulations to process.}

\item{list_simuM}{The list of simulations.}

\item{u}{The quantile value.}

\item{df_lags}{The dataframe with spatial and temporal lag values.}

\item{t0}{The starting time.}

\item{true_param}{The true variogram parameter (beta1, beta2, alpha1, alpha2)}

\item{hmax}{The maximum spatial lag value. Default is NA.}
}
\value{
The optimized parameters.
}
\description{
This function processes the simulation data and optimizes the parameters
using the negative log-likelihood function.
}
