% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{save_results_optim}
\alias{save_results_optim}
\title{save_results_optim function}
\usage{
save_results_optim(result, filename)
}
\arguments{
\item{result}{The result of the optimization process.}

\item{filename}{The name of the file to save the results.}
}
\description{
Save the results of the optimization process in a CSV file.
}
