% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{select_extreme_episodes}
\alias{select_extreme_episodes}
\title{select_extreme_episodes function}
\usage{
select_extreme_episodes(
  sites_coords,
  data,
  quantile,
  min_spatial_dist,
  min_time_dist,
  delta,
  n_max_episodes,
  beta = 0
)
}
\arguments{
\item{sites_coords}{The coordinates of the sites.}

\item{data}{The data.}

\item{quantile}{The quantile value.}

\item{min_spatial_dist}{The minimum spatial distance.}

\item{min_time_dist}{The minimum time distance.}

\item{delta}{The episode length.}

\item{n_max_episodes}{The maximum number of episodes to select.}

\item{beta}{The buffer size. Default is 0.}
}
\value{
The selected points.
}
\description{
Select extreme episodes based on the data, the quantile, the minimum spatial
distance, the minimum time distance and the episode length.
}
