% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{select_extreme_episodes}
\alias{select_extreme_episodes}
\title{select_extreme_episodes function}
\usage{
select_extreme_episodes(
  sites_coords,
  data,
  min_spatial_dist,
  episode_size,
  n_max_episodes = 10000,
  time_ext = 0,
  latlon = TRUE,
  quantile = NULL,
  threshold = NULL
)
}
\arguments{
\item{sites_coords}{The coordinates of the sites.}

\item{data}{The rainfall dataframe.}

\item{min_spatial_dist}{The minimum spatial distance between two episodes.}

\item{episode_size}{The temporal window size, i.e. the size of the episode.}

\item{n_max_episodes}{The maximum number of episodes to select. Default is 10000.}

\item{time_ext}{The temporal window extension. Default is 0.}

\item{latlon}{A boolean value to indicate if the locations are in latitude and longitude. Default is TRUE.}

\item{quantile}{The quantile value.}
}
\value{
A list of selected points (s0, t0) and the corresponding threshold value u_s0 for each episode.
}
\description{
This function selects extreme episodes based on the given data and a 
quantile. A spatio-temporal specific neighborhood is considered to avoid
selecting nearby episodes. Each episode is selected according to a 
r-Pareto process with r(X) = X(s0, t0). Each episode will have a size of
episode_size.
}
