% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_rpareto}
\alias{sim_rpareto}
\title{sim_rpareto function}
\usage{
sim_rpareto(
  beta1,
  beta2,
  alpha1,
  alpha2,
  x,
  y,
  t,
  adv = c(0, 0),
  t0 = 0,
  nres = 1,
  random_s0 = FALSE,
  s0 = c(1, 1)
)
}
\arguments{
\item{beta1}{The value of beta1.}

\item{beta2}{The value of beta2.}

\item{alpha1}{The value of alpha1.}

\item{alpha2}{The value of alpha2.}

\item{x}{Vector for the first dimension (spatial x in our case).}

\item{y}{Vector for the second dimension (spatial y in our case)}

\item{t}{Vector for the third dimension (time in our case).}

\item{adv}{The advection coordinates vector. Default is c(0, 0).}

\item{t0}{Conditional temporal time. Default is 1.}

\item{nres}{The number of simulations to perform. Default is 1.}

\item{random_s0}{Logical value indicating whether to choose a random s0.
Default is FALSE.}

\item{s0}{Vector of dimension 2 for the spatial conditioning point.}
}
\value{
The result of the simulation.
}
\description{
This function performs a simulation of a spatio-temporal r-Pareto process
using a fractionnal Brownian motion model and based on the David Leber code.
}
