% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{simulate_excesses}
\alias{simulate_excesses}
\title{Simulate excesses}
\usage{
simulate_excesses(Tmax, tau, h_vect, chi)
}
\arguments{
\item{Tmax}{The maximum time observation.}

\item{tau}{The temporal lag vector.}

\item{h_vect}{The spatial lag vector.}

\item{chi}{The spatio-temporal extremogram matrix.}

\item{df_dist}{The distances long dataframe.}
}
\value{
The simulated excesses matrix and the chi vector.
}
\description{
This function simulates excesses following a binomial distribution with given
probability parameters equal to the spatio-temporal extremogram values.
}
