% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{spatial_chi}
\alias{spatial_chi}
\title{Function to estimate the spatial extremogram}
\usage{
spatial_chi(lags, rad_mat, data_rain, quantile, zeros = TRUE, mid = TRUE)
}
\arguments{
\item{lags}{The lags vector.}

\item{rad_mat}{The matrix of radii.}

\item{data_rain}{The rainfall data.}

\item{quantile}{The quantile value.}

\item{zeros}{Logical value indicating whether to include zero values in the
calculation.}

\item{mid}{Logical value indicating whether to return the mean values for
the plot.}
}
\value{
The spatial extremogram.
}
\description{
This function calculates the spatial extremogram for a given lags vector and
and radius matrix and rainfall data.
}
