% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{spatial_chi_alldist}
\alias{spatial_chi_alldist}
\title{Calculate spatial chi-squared distances}
\usage{
spatial_chi_alldist(df_dist, data_rain, quantile, hmax = NA, comephore = FALSE)
}
\arguments{
\item{df_dist}{The distance dataframe.}

\item{data_rain}{The rainfall data.}

\item{quantile}{The quantile value.}

\item{hmax}{The maximum spatial lag value (optional).}

\item{comephore}{Logical value indicating whether we work on the COMEPHORE 
dataset. Default is FALSE.}
}
\value{
The spatial chi-squared distances.
}
\description{
This function calculates the spatial extremogram for a given
distance matrix and rainfall data.
}
\examples{
df_dist <- matrix(c(0, 1, 2, 1, 0, 3, 2, 3, 0), nrow = 3)
data_rain <- c(10, 20, 30)
quantile <- 0.5
spatial_chi_alldist(df_dist, data_rain, quantile)

}
