% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{theoretical_chi}
\alias{theoretical_chi}
\title{Compute the theoretical chi dataframe.}
\usage{
theoretical_chi(
  params,
  df_lags,
  latlon = FALSE,
  wind_vect = NA,
  directional = TRUE
)
}
\arguments{
\item{params}{A vector of variogram parameters.}

\item{df_lags}{A dataframe with spatial and temporal lag values.}

\item{latlon}{A boolean value to indicate if the locations are in latitude
and longitude. Default is FALSE.}

\item{wind_vect}{A vector of wind values. Default is NA.}

\item{directional}{A boolean value to indicate if the variogram is
directional. Default is TRUE.}
}
\value{
The theoretical chi
}
\description{
This function calculates the theoretical chi dataframe, based on the given
variogram parameters.
}
