% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{theorical_chi_mat}
\alias{theorical_chi_mat}
\title{Compute the theoretical chi matrix.}
\usage{
theorical_chi_mat(params, h_vect, tau)
}
\arguments{
\item{params}{A vector of parameters.}

\item{h_vect}{A vector of spatial lag values.}

\item{tau}{A vector of temporal lag values.}
}
\value{
The theoretical chi matrix.
}
\description{
This function calculates the theoretical chi matrix based on the given
parameters.
}
