% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{vario}
\alias{vario}
\title{Calculate the spatial or temporal variogram}
\usage{
vario(x, c, alpha)
}
\arguments{
\item{x}{The corresponding lags.}

\item{c}{The c variogram parameter. (log(beta))}

\item{alpha}{The alpha variogram parameter.}
}
\value{
The calculated spatial or temporal variogram.
}
\description{
This function calculates the spatial or temporal variogram based on the given
WLSE estimated parameters c and alpha. (c = log(beta))
}
\examples{
x <- c(1, 2, 3, 4, 5)
c <- 1
alpha <- 0.5
vario <- vario(x, c, alpha)
print(vario)
# Output: [1] 2.718282 7.389056 20.085537 54.598150 148.413159
}
