% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wlse_tablatex}
\alias{wlse_tablatex}
\title{wlse_tablatex function}
\usage{
wlse_tablatex(summary_df, ind, filename = "")
}
\arguments{
\item{summary_df}{The summary dataframe containing the data to be converted
into a table.}

\item{ind}{The index of the summary dataframe to be used for generating the
table.}

\item{filename}{The name of the file to save the generated LaTeX table.
If not provided, the table will be printed to the console.}
}
\value{
The generated LaTeX table.
}
\description{
This function generates a LaTeX table from a summary dataframe of WLSE
results.
}
\examples{
# Generate a summary dataframe
summary_df <- data.frame(
  Group = c("A", "B", "C"),
  Mean = c(10, 20, 30),
  SD = c(1, 2, 3)
)

# Generate LaTeX table from the summary dataframe
wlse_tablatex(summary_df, 1)

}
